Update planck_param_sims.md
I could swear that i saw repeats of phrases in the original text that was in the article, but i'm not finding them in the original TEXT. That leads me to believe that possibly i am an idiot, or that the software which writes the final from the text is somehow faulty. Anyway, here 'tis. -- john